Modularized entire code base
