+++ include/LED_Display.h
@@ -1,3 +1,12 @@
+/*
+ ******************************************************************************
+ * File Name          : LED_Display.h
+ * Description        : This file contains all the function prototypes for
+ *                      the LED_Display.c file
+ ****************************************************************************** */
+
+/* Define to prevent recursive inclusion -------------------------------------*/
+
 #ifndef LEDDISPLAY_H
 #define LEDDISPLAY_H
 
@@ -9,5 +18,6 @@ uint8_t dotRegister[LEDDISPLAY_MAXCHARS * 5];
 
 /* Loads bits into Dot Register */
 void fillDotRegister(char whatCharacter, uint8_t whatPosition);
+void clearDotRegister(void);
 
 #endif
