+++ src/main.c
@@ -38,54 +38,20 @@
 
 /* Includes ------------------------------------------------------------------*/
 #include "main.h"
-#include "defines.h"
 #include "stm32f4xx_hal.h"
+#include "gpio.h"
+#include "defines.h"
+#include "adc.h"
+#include "dma.h"
 #include "LED_Display.h"
-
-/* Calibration for the ADC sensors, lower value -> see further */
-/*                                  higher value -> see closer */
-/* R_ADC is +10 more then L_ADC
- * LF_ADC is +10 more then RF_ADC
- */
-#define L_ADC 30       // 0
-#define LF_ADC 60      // 1
-#define RF_ADC 50      // 2
-#define R_ADC  40      // 3
-
-/* Private variables ---------------------------------------------------------*/
-ADC_HandleTypeDef hadc1;
-
-TIM_HandleTypeDef htim2;
-TIM_HandleTypeDef htim4;
-TIM_HandleTypeDef htim5;
-SPI_HandleTypeDef hspi3;
-
-
-/* ADC DMA buffers */
-DMA_HandleTypeDef g_DmaHandle;
-enum { ADC_BUFFER_LENGTH = 8192 };
-uint32_t g_ADCBuffer[ADC_BUFFER_LENGTH];
-uint32_t IR_values[4];
+#include "motors.h"
+#include "spi.h"
+#include "tim.h"
 
 /* Private function prototypes -----------------------------------------------*/
 void SystemClock_Config(void);
-static void MX_GPIO_Init(void);
-static void MX_ADC1_Init(void);
-static void MX_TIM5_Init(void);
-static void MX_TIM2_Init(void);
-static void MX_TIM4_Init(void);
 
-static void SPI3_Init(void);
-
-void HAL_TIM_MspPostInit(TIM_HandleTypeDef *htim);
-void ConfigureDMA();
-
-void user_pwm_setValue_LeftMotors(uint16_t value);
-void user_pwm_setValue_RightMotors(uint16_t value);
-
-/* Function to print out ADC value reading to the LEDs */
-static void ADC_LED_DEBUG(uint32_t value);
-
+/* Main program */
 int main(void)
 {
 	/* Reset of all peripherals, Initializes the Flash interface and the Systick. */
@@ -97,214 +63,26 @@ int main(void)
 	/* Initialize all configured peripherals */
 	MX_GPIO_Init();
 	MX_ADC1_Init();
-	MX_TIM5_Init();
 	MX_TIM2_Init();
 	MX_TIM4_Init();
-	SPI3_Init();
+	MX_TIM5_Init();
+	MX_SPI3_Init();
+	MX_DMA_Init();
 
-	/* USER CODE BEGIN 2 */
+	/* Enable Emitter pins when mouse powers on */
+	emitter_Init();
+
 	/*
-  HAL_TIM_Base_Start(&htim4);
   HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_1);
   HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_2);
   HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_3);
   HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_4);
-
-
-
 	 */
-	/* Set GPIO pins, GPIO_PIN_SET turns on pin while GPIO_PIN_RESET turns off */
-	/* Can set LED pins to clear initially */
-	/* Enable Emitter pins when mouse powers on */
-//	HAL_GPIO_WritePin(GPIOA, GPIO_PIN_4, GPIO_PIN_RESET);  //LED
-//	HAL_GPIO_WritePin(GPIOA, GPIO_PIN_5, GPIO_PIN_RESET);  //LED
-//	HAL_GPIO_WritePin(GPIOA, GPIO_PIN_7, GPIO_PIN_RESET);  //LED
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_4, GPIO_PIN_RESET);  //LED
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_5, GPIO_PIN_RESET);  //LED
-//	HAL_GPIO_WritePin(GPIOB, GPIO_PIN_12, GPIO_PIN_RESET); //LED
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_6, GPIO_PIN_RESET);  //LED
-	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_7, GPIO_PIN_SET);  //L_EMITTER
-	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_8, GPIO_PIN_SET);  //R_EMITTER
-	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_9, GPIO_PIN_SET);  //LF_EMITTER
-	HAL_GPIO_WritePin(GPIOA, GPIO_PIN_8, GPIO_PIN_SET);  //RF_EMITTER
-//	HAL_GPIO_WritePin(GPIOA, GPIO_PIN_11, GPIO_PIN_RESET); //Buzzer
-
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_14, GPIO_PIN_RESET); //LED
-
-
-//	uint8_t test[] = {65,66,67,68};
-
-	/* LCD display start */
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_11, GPIO_PIN_RESET);  //RS
-//	HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_RESET);  //CE LOW
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_12, GPIO_PIN_SET);  //DATA_IN
-
-	// EACH DIGIT IS 40 BITS
-//	fillDotRegister('A', 1);
-//	HAL_SPI_Transmit(&hspi3, dotRegister, 40, 0xFF);
-//	while(HAL_SPI_GetState(&hspi3) == HAL_SPI_STATE_BUSY);
-
-//	uint8_t test[] = {0x7F, 0x04, 0x08, 0x10, 0x7F};
-//	uint8_t test[] = {1, 1, 1, 1};
-//	uint8_t message[] = {"0x7F", "0x41", "0x41", "0x22", "0x1C"};
-//	uint8_t message[] = {0x7F, 0x41, 0x41, 0x22, 0x1C};
-
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_11, GPIO_PIN_RESET);  //RS
-//	HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_RESET);  //CE LOW
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_12, GPIO_PIN_SET);  //DATA_IN
-//
-//
-//    clearDotRegister();
-//    fillDotRegister('A', 1);
-
-////	HAL_SPI_Transmit(&hspi3, dotRegister, 160, 0xFF);
-//
-//
-//	HAL_SPI_Transmit(&hspi3, dotRegister, 1280, 0xFF);
-//	HAL_Delay(10);
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_10, GPIO_PIN_SET);  //CLK
-//
-//	HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_SET);  //CE
-
-
-
-
-
-//
-//		HAL_GPIO_WritePin(GPIOC, GPIO_PIN_11, GPIO_PIN_RESET);  //RS
-//		HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_RESET);  //CE LOW
-//		HAL_GPIO_WritePin(GPIOC, GPIO_PIN_12, GPIO_PIN_SET);  //DATA_IN
-//
-//		HAL_SPI_Transmit(&hspi3, dotRegister, 1280, HAL_MAX_DELAY);
-//		while(HAL_SPI_GetState(&hspi3) == HAL_SPI_STATE_BUSY);
-//
-//		HAL_Delay(10);
-//		HAL_GPIO_WritePin(GPIOC, GPIO_PIN_10, GPIO_PIN_SET);  //CLK
-//
-//		HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_SET);  //CE
-
-//	fillDotRegister('B', 2);
-//	HAL_SPI_Transmit(&hspi3, dotRegister, 40, 0xFF);
-//	while(HAL_SPI_GetState(&hspi3) == HAL_SPI_STATE_BUSY);
-//
-//	fillDotRegister('C', 3);
-//	HAL_SPI_Transmit(&hspi3, dotRegister, 40, 0xFF);
-//	while(HAL_SPI_GetState(&hspi3) == HAL_SPI_STATE_BUSY);
-//
-//	fillDotRegister('D', 4);
-//	HAL_SPI_Transmit(&hspi3, dotRegister, 40, 0xFF);
-//	while(HAL_SPI_GetState(&hspi3) == HAL_SPI_STATE_BUSY);
-//
-
-
-
-//
-//	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_10, GPIO_PIN_SET);  //CLK
-//
-//	HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_SET);  //CE
-
-	/* LCD display end */
-
-	/* ADC */
-
-	ConfigureDMA();
-	HAL_ADC_Start_DMA(&hadc1, g_ADCBuffer, ADC_BUFFER_LENGTH);
 
 	while (1)
 	{
-		if (IR_values[0] >= L_ADC)
-		{
-			HAL_GPIO_WritePin(GPIOC, GPIO_PIN_5, GPIO_PIN_SET);  //LED
-		}
-		else
-		{
-			HAL_GPIO_WritePin(GPIOC, GPIO_PIN_5, GPIO_PIN_RESET);  //LED
-		}
-
-		if (IR_values[1] >= LF_ADC)
-		{
-			HAL_GPIO_WritePin(GPIOC, GPIO_PIN_4, GPIO_PIN_SET);  //LED
-		}
-		else
-		{
-			HAL_GPIO_WritePin(GPIOC, GPIO_PIN_4, GPIO_PIN_RESET);  //LED
-		}
-		if (IR_values[2] >= RF_ADC)
-		{
-			HAL_GPIO_WritePin(GPIOA, GPIO_PIN_7, GPIO_PIN_SET);  //LED
-		}
-		else
-		{
-			HAL_GPIO_WritePin(GPIOA, GPIO_PIN_7, GPIO_PIN_RESET);  //LED
-		}
-		if (IR_values[3] >= R_ADC)
-		{
-			HAL_GPIO_WritePin(GPIOA, GPIO_PIN_5, GPIO_PIN_SET);  //LED
-		}
-		else
-		{
-			HAL_GPIO_WritePin(GPIOA, GPIO_PIN_5, GPIO_PIN_RESET);  //LED
-		}
+		ADC_LED_Distance_Tester();
 	}
-}
-
-void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef* AdcHandle)
-{
-	int i,j,l=0,lf=0,rf=0,r=0;
-	for(i = 0; i < ADC_BUFFER_LENGTH; i++){
-		j = i % 4;
-		switch(j) {
-		case 0:
-			l += g_ADCBuffer[i];
-			break;
-		case 1:
-			lf += g_ADCBuffer[i];
-			break;
-		case 2:
-			rf += g_ADCBuffer[i];
-			break;
-		case 3:
-			r += g_ADCBuffer[i];
-			break;
-		}
-	}
-	IR_values[0] = l/2048;
-	IR_values[1] = lf/2048;
-	IR_values[2] = rf/2048;
-	IR_values[3] = r/2048;
-}
-
-void DMA2_Stream4_IRQHandler()
-{
-	HAL_DMA_IRQHandler(&g_DmaHandle);
-}
-
-void ADC_IRQHandler()
-{
-	HAL_ADC_IRQHandler(&hadc1);
-}
-
-/* Init SPI3 */
-static void SPI3_Init(void)
-{
-	hspi3.Instance = SPI3;
-	hspi3.Init.Mode = SPI_MODE_MASTER;
-	hspi3.Init.Direction = SPI_DIRECTION_2LINES;
-	hspi3.Init.DataSize = SPI_DATASIZE_8BIT;
-	hspi3.Init.CLKPolarity = SPI_POLARITY_LOW;
-	hspi3.Init.CLKPhase = SPI_PHASE_1EDGE;
-	hspi3.Init.NSS = SPI_NSS_SOFT;
-	hspi3.Init.BaudRatePrescaler = SPI_BAUDRATEPRESCALER_128;
-	hspi3.Init.FirstBit = SPI_FIRSTBIT_MSB;
-	hspi3.Init.TIMode = SPI_TIMODE_DISABLE;
-	hspi3.Init.CRCCalculation = SPI_CRCCALCULATION_DISABLE;
-	hspi3.Init.CRCPolynomial = 10;
-	if (HAL_SPI_Init(&hspi3) != HAL_OK)
-	{
-		Error_Handler();
-	}
-	__SPI3_CLK_ENABLE();
-	__HAL_SPI_ENABLE(&hspi3);
 }
 
 /** System Clock Configuration
@@ -358,330 +136,6 @@ void SystemClock_Config(void)
 	HAL_NVIC_SetPriority(SysTick_IRQn, 0, 0);
 }
 
-/* ADC1 init function */
-static void MX_ADC1_Init(void)
-{
-	HAL_NVIC_SetPriority(ADC_IRQn, 0, 0);
-	HAL_NVIC_EnableIRQ(ADC_IRQn);
-
-	ADC_ChannelConfTypeDef sConfig;
-
-	/**Configure the global features of the ADC (Clock, Resolution, Data Alignment and number of conversion)
-	 */
-	hadc1.Instance = ADC1;
-	hadc1.Init.ClockPrescaler = ADC_CLOCK_SYNC_PCLK_DIV2;
-	hadc1.Init.Resolution = ADC_RESOLUTION_12B;
-	hadc1.Init.ScanConvMode = ENABLE;
-	hadc1.Init.ContinuousConvMode = ENABLE;
-	hadc1.Init.DiscontinuousConvMode = DISABLE;
-	hadc1.Init.NbrOfDiscConversion = 0;
-	hadc1.Init.ExternalTrigConvEdge = ADC_EXTERNALTRIGCONVEDGE_NONE;
-	hadc1.Init.ExternalTrigConv = ADC_EXTERNALTRIGCONV_T1_CC1;
-	hadc1.Init.DataAlign = ADC_DATAALIGN_RIGHT;
-	hadc1.Init.NbrOfConversion = 4;
-	hadc1.Init.DMAContinuousRequests = ENABLE;
-	hadc1.Init.EOCSelection = DISABLE;
-	if (HAL_ADC_Init(&hadc1) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	/**Configure for the selected ADC regular channel its corresponding rank in the sequencer and its sample time.
-	 */
-	sConfig.Channel = ADC_CHANNEL_8;
-	sConfig.Rank = 1;
-	sConfig.SamplingTime = ADC_SAMPLETIME_56CYCLES;
-	if (HAL_ADC_ConfigChannel(&hadc1, &sConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	sConfig.Channel = ADC_CHANNEL_9;
-	sConfig.Rank = 2;
-	sConfig.SamplingTime = ADC_SAMPLETIME_56CYCLES;
-	if (HAL_ADC_ConfigChannel(&hadc1, &sConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	sConfig.Channel = ADC_CHANNEL_2;
-	sConfig.Rank = 3;
-	sConfig.SamplingTime = ADC_SAMPLETIME_56CYCLES;
-	if (HAL_ADC_ConfigChannel(&hadc1, &sConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	sConfig.Channel = ADC_CHANNEL_3;
-	sConfig.Rank = 4;
-	sConfig.SamplingTime = ADC_SAMPLETIME_56CYCLES;
-	if (HAL_ADC_ConfigChannel(&hadc1, &sConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-
-}
-
-/* TIM2 init function */
-static void MX_TIM2_Init(void)
-{
-
-	TIM_Encoder_InitTypeDef sConfig;
-	TIM_MasterConfigTypeDef sMasterConfig;
-
-	htim2.Instance = TIM2;
-	htim2.Init.Prescaler = 0;
-	htim2.Init.CounterMode = TIM_COUNTERMODE_UP;
-	htim2.Init.Period = 0;
-	htim2.Init.ClockDivision = TIM_CLOCKDIVISION_DIV1;
-	sConfig.EncoderMode = TIM_ENCODERMODE_TI1;
-	sConfig.IC1Polarity = TIM_ICPOLARITY_RISING;
-	sConfig.IC1Selection = TIM_ICSELECTION_DIRECTTI;
-	sConfig.IC1Prescaler = TIM_ICPSC_DIV1;
-	sConfig.IC1Filter = 0;
-	sConfig.IC2Polarity = TIM_ICPOLARITY_RISING;
-	sConfig.IC2Selection = TIM_ICSELECTION_DIRECTTI;
-	sConfig.IC2Prescaler = TIM_ICPSC_DIV1;
-	sConfig.IC2Filter = 0;
-	if (HAL_TIM_Encoder_Init(&htim2, &sConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	sMasterConfig.MasterOutputTrigger = TIM_TRGO_RESET;
-	sMasterConfig.MasterSlaveMode = TIM_MASTERSLAVEMODE_DISABLE;
-	if (HAL_TIMEx_MasterConfigSynchronization(&htim2, &sMasterConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-}
-
-/* TIM4 init function */
-static void MX_TIM4_Init(void)
-{
-
-	TIM_MasterConfigTypeDef sMasterConfig;
-	TIM_OC_InitTypeDef sConfigOC;
-
-	htim4.Instance = TIM4;
-	htim4.Init.Prescaler = 0;
-	htim4.Init.CounterMode = TIM_COUNTERMODE_UP;
-	htim4.Init.Period = 0;
-	htim4.Init.ClockDivision = TIM_CLOCKDIVISION_DIV1;
-	if (HAL_TIM_PWM_Init(&htim4) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	sMasterConfig.MasterOutputTrigger = TIM_TRGO_RESET;
-	sMasterConfig.MasterSlaveMode = TIM_MASTERSLAVEMODE_DISABLE;
-	if (HAL_TIMEx_MasterConfigSynchronization(&htim4, &sMasterConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	sConfigOC.OCMode = TIM_OCMODE_PWM1;
-	sConfigOC.Pulse = 0;
-	sConfigOC.OCPolarity = TIM_OCPOLARITY_HIGH;
-	sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_1) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_2) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_3) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_4) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	HAL_TIM_MspPostInit(&htim4);
-
-}
-
-/* TIM5 init function */
-static void MX_TIM5_Init(void)
-{
-
-	TIM_Encoder_InitTypeDef sConfig;
-	TIM_MasterConfigTypeDef sMasterConfig;
-
-	htim5.Instance = TIM5;
-	htim5.Init.Prescaler = 0;
-	htim5.Init.CounterMode = TIM_COUNTERMODE_UP;
-	htim5.Init.Period = 0;
-	htim5.Init.ClockDivision = TIM_CLOCKDIVISION_DIV1;
-	sConfig.EncoderMode = TIM_ENCODERMODE_TI1;
-	sConfig.IC1Polarity = TIM_ICPOLARITY_RISING;
-	sConfig.IC1Selection = TIM_ICSELECTION_DIRECTTI;
-	sConfig.IC1Prescaler = TIM_ICPSC_DIV1;
-	sConfig.IC1Filter = 0;
-	sConfig.IC2Polarity = TIM_ICPOLARITY_RISING;
-	sConfig.IC2Selection = TIM_ICSELECTION_DIRECTTI;
-	sConfig.IC2Prescaler = TIM_ICPSC_DIV1;
-	sConfig.IC2Filter = 0;
-	if (HAL_TIM_Encoder_Init(&htim5, &sConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-	sMasterConfig.MasterOutputTrigger = TIM_TRGO_RESET;
-	sMasterConfig.MasterSlaveMode = TIM_MASTERSLAVEMODE_DISABLE;
-	if (HAL_TIMEx_MasterConfigSynchronization(&htim5, &sMasterConfig) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-
-}
-
-/** Configure pins as
- * Analog
- * Input
- * Output
- * EVENT_OUT
- * EXTI
-     PA9   ------> USART1_TX
-     PA10   ------> USART1_RX
-     PC10   ------> SPI3_SCK
-     PC11   ------> SPI3_MISO
-     PC12   ------> SPI3_MOSI
- */
-static void MX_GPIO_Init(void)
-{
-
-	GPIO_InitTypeDef GPIO_InitStruct;
-
-	/* GPIO Ports Clock Enable */
-	__HAL_RCC_GPIOC_CLK_ENABLE();
-	__HAL_RCC_GPIOA_CLK_ENABLE();
-	__HAL_RCC_GPIOB_CLK_ENABLE();
-	__HAL_RCC_GPIOD_CLK_ENABLE();
-
-	/*Configure GPIO pin Output Level */
-	HAL_GPIO_WritePin(GPIOC, GPIO_PIN_14|GPIO_PIN_4|GPIO_PIN_5|GPIO_PIN_6
-			|GPIO_PIN_7|GPIO_PIN_8|GPIO_PIN_9, GPIO_PIN_RESET);
-
-	/*Configure GPIO pin Output Level */
-	HAL_GPIO_WritePin(GPIOA, GPIO_PIN_4|GPIO_PIN_5|GPIO_PIN_7|GPIO_PIN_8
-			|GPIO_PIN_11, GPIO_PIN_RESET);
-
-	/*Configure GPIO pin Output Level */
-	HAL_GPIO_WritePin(GPIOB, GPIO_PIN_12, GPIO_PIN_RESET);
-
-	/*Configure GPIO pin Output Level */
-	HAL_GPIO_WritePin(GPIOD, GPIO_PIN_2, GPIO_PIN_RESET);
-
-	/*Configure GPIO pins : PC14 PC4 PC5 PC6
-                           PC7 PC8 PC9 */
-	GPIO_InitStruct.Pin = GPIO_PIN_14|GPIO_PIN_4|GPIO_PIN_5|GPIO_PIN_6
-			|GPIO_PIN_7|GPIO_PIN_8|GPIO_PIN_9;
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-	GPIO_InitStruct.Pull = GPIO_NOPULL;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-	HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-	/*Configure GPIO pins : PA4 PA5 PA7 PA8
-                           PA11 */
-	GPIO_InitStruct.Pin = GPIO_PIN_4|GPIO_PIN_5|GPIO_PIN_7|GPIO_PIN_8
-			|GPIO_PIN_11;
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-	GPIO_InitStruct.Pull = GPIO_NOPULL;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-	HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-	/*Configure GPIO pin : PB12 */
-	GPIO_InitStruct.Pin = GPIO_PIN_12;
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-	GPIO_InitStruct.Pull = GPIO_NOPULL;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-	HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-	/*Configure GPIO pins : PA9 PA10 */
-	GPIO_InitStruct.Pin = GPIO_PIN_9|GPIO_PIN_10;
-	GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-	GPIO_InitStruct.Pull = GPIO_NOPULL;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-	GPIO_InitStruct.Alternate = GPIO_AF7_USART1;
-	HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-	/* SPI GPIO INIT START */
-    __HAL_RCC_SPI3_CLK_ENABLE();
-    /**SPI3 GPIO Configuration
-        PC10     ------> SPI3_SCK
-        PC11     ------> SPI3_MISO
-        PC12     ------> SPI3_MOSI
-        */
-    GPIO_InitStruct.Pin = GPIO_PIN_10|GPIO_PIN_11|GPIO_PIN_12;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF6_SPI3;
-    HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-	/*Configure GPIO pin : PD2 */
-    /* SPI LCD Display CE pin */
-	GPIO_InitStruct.Pin = GPIO_PIN_2;
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-	GPIO_InitStruct.Pull = GPIO_NOPULL;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-	HAL_GPIO_Init(GPIOD, &GPIO_InitStruct);
-
-	/* SPI GPIO INIT END */
-}
-
-/* USER CODE BEGIN 4 */
-/* Pulse PWM value change for left motors (CH1/CH2) */
-void user_pwm_setValue_LeftMotors(uint16_t value)
-{
-	TIM_OC_InitTypeDef sConfigOC;
-
-	sConfigOC.OCMode = TIM_OCMODE_PWM1;
-	sConfigOC.Pulse = value;
-	sConfigOC.OCPolarity = TIM_OCPOLARITY_HIGH;
-	sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_1) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_2) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_1);
-	HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_2);
-}
-
-/* Pulse PWM value change for right motors (CH3/CH4) */
-void user_pwm_setValue_RightMotors(uint16_t value)
-{
-	TIM_OC_InitTypeDef sConfigOC;
-
-	sConfigOC.OCMode = TIM_OCMODE_PWM1;
-	sConfigOC.Pulse = value;
-	sConfigOC.OCPolarity = TIM_OCPOLARITY_HIGH;
-	sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_3) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	if (HAL_TIM_PWM_ConfigChannel(&htim4, &sConfigOC, TIM_CHANNEL_4) != HAL_OK)
-	{
-		_Error_Handler(__FILE__, __LINE__);
-	}
-	HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_3);
-	HAL_TIM_PWM_Start(&htim4, TIM_CHANNEL_4);
-}
-
-/* USER CODE END 4 */
-
 /**
  * @brief  This function is executed in case of error occurrence.
  * @param  None
@@ -714,58 +168,6 @@ void assert_failed(uint8_t* file, uint32_t line)
 	/* USER CODE END 6 */
 
 }
-
-void ConfigureDMA(){
-	__DMA2_CLK_ENABLE();
-	g_DmaHandle.Instance = DMA2_Stream4;
-
-	g_DmaHandle.Init.Channel  = DMA_CHANNEL_0;
-	g_DmaHandle.Init.Direction = DMA_PERIPH_TO_MEMORY;
-	g_DmaHandle.Init.PeriphInc = DMA_PINC_DISABLE;
-	g_DmaHandle.Init.MemInc = DMA_MINC_ENABLE;
-	g_DmaHandle.Init.PeriphDataAlignment = DMA_PDATAALIGN_WORD;
-	g_DmaHandle.Init.MemDataAlignment = DMA_MDATAALIGN_WORD;
-	g_DmaHandle.Init.Mode = DMA_CIRCULAR;
-	g_DmaHandle.Init.Priority = DMA_PRIORITY_HIGH;
-	g_DmaHandle.Init.FIFOMode = DMA_FIFOMODE_DISABLE;
-	g_DmaHandle.Init.FIFOThreshold = DMA_FIFO_THRESHOLD_HALFFULL;
-	g_DmaHandle.Init.MemBurst = DMA_MBURST_SINGLE;
-	g_DmaHandle.Init.PeriphBurst = DMA_PBURST_SINGLE;
-
-	HAL_DMA_Init(&g_DmaHandle);
-
-	__HAL_LINKDMA(&hadc1, DMA_Handle, g_DmaHandle);
-
-	HAL_NVIC_SetPriority(DMA2_Stream4_IRQn, 0, 0);
-	HAL_NVIC_EnableIRQ(DMA2_Stream4_IRQn);
-}
-
-/* Function to print out ADC value reading to the LEDs */
-static void ADC_LED_DEBUG(uint32_t value) {
-	value = HAL_ADC_GetValue(&hadc1);
-	int ones = value % 10;
-	int tens = (value % 100)/10;
-	int hundreds = (value % 1000)/100;
-	int thousands = value/1000;
-	for(;;){
-		if(thousands>0) HAL_GPIO_WritePin(GPIOC, GPIO_PIN_5, GPIO_PIN_SET);
-		if(hundreds>0) HAL_GPIO_WritePin(GPIOC, GPIO_PIN_4, GPIO_PIN_SET);
-		if(tens>0) HAL_GPIO_WritePin(GPIOA, GPIO_PIN_7, GPIO_PIN_SET);
-		if(ones>0) HAL_GPIO_WritePin(GPIOA, GPIO_PIN_5, GPIO_PIN_SET);
-		HAL_Delay(250);
-		thousands--;
-		hundreds--;
-		tens--;
-		ones--;
-		HAL_GPIO_WritePin(GPIOC, GPIO_PIN_5, GPIO_PIN_RESET);
-		HAL_GPIO_WritePin(GPIOC, GPIO_PIN_4, GPIO_PIN_RESET);
-		HAL_GPIO_WritePin(GPIOA, GPIO_PIN_7, GPIO_PIN_RESET);
-		HAL_GPIO_WritePin(GPIOA, GPIO_PIN_5, GPIO_PIN_RESET);
-		HAL_Delay(250);
-		if(thousands<0 && tens<0 && hundreds<0 && ones<0) break;
-	}
-}
-
 
 #endif
